+++ .gitignore
@@ -1,2 +1,3 @@
 /venv/
 .idea
+tgsession.session
