+++ config/config.ini
@@ -1,12 +1,12 @@
 [telegram]
-api_id =
-api_hash =
+api_id = EnterIdHere
+api_hash = EnterHashHere
 sessionfile = tgsession
 chatroom = 3C Quick Stats
 
 [commas]
-key =
-secret =
+key = EnterKeyHere
+secret = EnterSecretHere
 
 [dcabot]
 prefix = 3CQSBOT
