+++ helpers/singlebot.py
@@ -1,6 +1,6 @@
 import re
 
-from signals import Signals
+from helpers.signals import Signals
 
 class SingleBot:
     def __init__(self, tg_data, bot_data, account_data, config, p3cw, logging):
